requirements.txt
streamlit==1.37.0
tensorflow==2.15.0
numpy==1.24.2
joblib==1.3.2
pillow==10.4.0
pandas==2.0.3
matplotlib==3.7.2